Tests/LibWeb/Text/expected/HTML/select-indexed-setter.txt

length: 1
text: new
